+++ lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:training_app/ui/home_page/home_page.dart';
+import 'package:training_app/ui/root_page/root_page.dart';
 
 void main() {
   runApp(const DearDiaryApp());
